+++ CHANGELOG.md
@@ -5,4 +5,6 @@ Changelog
 -----
 * Add LangFileCleaner class to sort and clean SugarCRM language files and make it easier for VCS
 * Add Installer class to extract and install sugar from a zip file.
+* Add SugarPDO to connect to sugarcrm database with php PDO
+* Add Metadata class to manage `field_meta_data` table.